avsafe_descriptors/pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "avsafe-descriptors"
version = "0.6.0"
description = "Privacy-preserving audiovisual metrology (LAeq, 1/3-octaves, IEEE-1789 flicker) with hash-chained reports and an HF-AVC corpus."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name="AV-SAFE Contributors"}]
dependencies = [
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "pydantic>=2.5",
  "PyYAML>=6.0",
  "Jinja2>=3.1",
  "sqlalchemy>=2.0",
  "pydantic-settings>=2.0",
  "python-dateutil>=2.9",
]

[project.optional-dependencies]
signing = ["pynacl>=1.5"]

[project.scripts]
avsafe-sim = "avsafe_descriptors.cli.sim:main"
avsafe-rules-run = "avsafe_descriptors.cli.rules_run:main"
avsafe-report = "avsafe_descriptors.cli.report:main"
hf-avc-ingest = "avsafe_descriptors.hf_avc.ingest_cli:main"

[tool.setuptools.packages.find]
where = ["."]
